Add funcs to create kernel tasks.
All tasks are in a circular-doubly linked list (through the next_task
and prev_task fields of task_struct).The run_queue is formed using the
next_rq and prev_rq fields of task_struct.
This was easier than having wrapper structs for task queues.

The kernel on startup is given a kernel_task task_struct, and a
kernel_mm mm_struct. Both are globals.